% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{assignConstraints}
\alias{assignConstraints}
\title{Applies suitable contraining functions to parameters.}
\usage{
assignConstraints(par, loc)
}
\arguments{
\item{par}{names vector with parameters}

\item{loc}{A data frame containing information on each involved parameter (list element
of objects of class \code{NAWRUmodel}, \code{TFPmodel}, \code{KuttnerModel}).}
}
\description{
Applies suitable contraining functions to parameters.
}
\keyword{internal}
