% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{autoNAWRUmodel}
\alias{autoNAWRUmodel}
\title{NAWRU model suggestion}
\usage{
autoNAWRUmodel(tsl, poss, nModels = 10)
}
\arguments{
\item{tsl}{A list of time series objects, see details.}

\item{poss}{A list with the characteristics of possible models (see \code{auto.gapProd)}}

\item{nModels}{Integer, the maximum number of models for each unobserved component model.}
}
\value{
A nested list with one model specification per list entry.
}
\description{
Finds the most suitable NAWRU models.
}
\keyword{internal}
