% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{autoTFPmodel}
\alias{autoTFPmodel}
\title{TFP model suggestion}
\usage{
autoTFPmodel(tsl, poss, nModels = 10)
}
\arguments{
\item{tsl}{A list of time series objects, see details.}

\item{poss}{A list with the characteristics of possible models (see \code{autoGapProd)}}

\item{nModels}{Integer, the maximum number of models for each unobserved component model.}
}
\value{
A nested list with one model specification per list entry.
}
\description{
Finds the most suitable TFP models.
}
\keyword{internal}
