% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{computeCovar}
\alias{computeCovar}
\title{Computes the covariance of the estimated parameters given restrictions.}
\usage{
computeCovar(par, loc, CV)
}
\arguments{
\item{par}{names vector with parameters}

\item{loc}{A data frame containing information on each involved parameter (list element
of objects of class \code{NAWRUmodel}, \code{TFPmodel}, \code{KuttnerModel}).}

\item{CV}{unconstrained variance covariance matrix}
}
\description{
Computes the covariance of the estimated parameters given restrictions.
}
\keyword{internal}
