% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraint}
\alias{constraint}
\title{Applies contraints to parameters.}
\usage{
constraint(par, type = NA)
}
\arguments{
\item{par}{names vector with parameters}

\item{type}{character with constraint type}
}
\description{
Applies contraints to parameters.
}
\keyword{internal}
