% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{cycleOptim}
\alias{cycleOptim}
\title{Find suitable cycle specification}
\usage{
cycleOptim(x, opt = c("AR1", "AR2", "RAR2"))
}
\arguments{
\item{x}{A time series.}

\item{opt}{A character vector with the cycle models to be tested. The default is 
\code{opt = c("AR1", "AR2", "RAR2")}.}
}
\value{
A character string with the chosen cycle model.
}
\description{
Finds the most suitable cycle model according to the BIC.
}
\keyword{internal}
