% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ts.R
\name{dateTsList}
\alias{dateTsList}
\title{Finds first/last starting/end date in list of time series deepening on the input functions.}
\usage{
dateTsList(x, FUN1 = start, FUN2 = max)
}
\arguments{
\item{x}{A list.}

\item{FUN1}{A function, namely either \code{start} or \code{end} from \code{stats}.}

\item{FUN2}{A function, namely either \code{max} or \code{min} from \code{base}.}
}
\description{
Finds first/last starting/end date in list of time series deepening on the input functions.
}
\keyword{internal}
