% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{Dconstraint}
\alias{Dconstraint}
\title{Extracts the derivative of the applied restriction function.}
\usage{
Dconstraint(par, type = NA)
}
\arguments{
\item{par}{names vector with parameters}

\item{type}{character with constraint type}
}
\description{
Extracts the derivative of the applied restriction function.
}
\keyword{internal}
