% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{.accessDfSystem}
\alias{.accessDfSystem}
\title{Accesses the internal data frame dfSystem which contains data on the parameters to be
estimated.}
\usage{
.accessDfSystem(model)
}
\arguments{
\item{model}{A model of class \code{NAWRUmodel} or \code{TFPmodel}.}
}
\value{
A data frame containing information on each involved parameter, for instance
  its corresponding system matrix, variable names, and parameter restrictions.
}
\description{
Accesses the internal data frame dfSystem which contains data on the parameters to be
estimated.
}
\keyword{internal}
