% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubs.R
\name{.aggregate}
\alias{.aggregate}
\title{Temporally Aggregates time series.}
\usage{
.aggregate(x, freqLow, FUN)
}
\arguments{
\item{x}{A time series object.}

\item{freqLow}{Frequency of low frequency series.}

\item{FUN}{A function for aggregation.}
}
\description{
Temporally Aggregates time series.
}
\keyword{internal}
