% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{.checkBoundaries}
\alias{.checkBoundaries}
\title{Checks whether estimated parameters lie on boundaries.}
\usage{
.checkBoundaries(fit, loc)
}
\arguments{
\item{fit}{Output of \code{fitSSM} from \code{KFAS}.}

\item{loc}{A data frame containing information on each involved parameter (list element
of objects of class \code{NAWRUmodel}, \code{TFPmodel}, \code{KuttnerModel}).}
}
\description{
Checks whether estimated parameters lie on boundaries.
}
\keyword{internal}
