% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.checkCubs}
\alias{.checkCubs}
\title{Checks the input variables for the procedure \code{cubs} for consistency and validity.}
\usage{
.checkCubs(tsCU, tsVA, lambda, frequency)
}
\arguments{
\item{tsCU}{A multiple time series containing three survey time series, the first list
element needs to be capacity utilization in industry, see details. Alternatively, a list
of time series can be supplied.}

\item{tsVA}{A multiple time series containing three value added series that correspond to
\code{tsCU}. Alternatively, a list of time series can be supplied.}

\item{lambda}{The smoothing parameter for the application of the HP filter (see details).
If not supplied, \code{lambda = 6.25} for yearly data and \code{lambda = 1600} for
quarterly data.}

\item{frequency}{The frequency of the computed cubs indicator. Possible entries are
\code{frequency = 1} (annual), \code{frequency = 4} (quarterly). The default is
\code{frequency = 1}.}
}
\description{
Checks the input variables for the procedure \code{cubs} for consistency and validity.
}
\keyword{internal}
