% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{.checkCV}
\alias{.checkCV}
\title{Checks the covariance matrix for invertibility and negative entries on the diagonal.}
\usage{
.checkCV(fit, loc)
}
\arguments{
\item{fit}{Output of \code{fitSSM} from \code{KFAS}.}

\item{loc}{A data frame containing information on each involved parameter (list element
of objects of class \code{NAWRUmodel}, \code{TFPmodel}, \code{KuttnerModel}).}
}
\description{
Checks the covariance matrix for invertibility and negative entries on the diagonal.
}
\keyword{internal}
