% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.checkModelMLEfit}
\alias{.checkModelMLEfit}
\title{Checks whether model, prior and MLE fit match.}
\usage{
.checkModelMLEfit(model, MLEfit)
}
\arguments{
\item{model}{An object of class TFPmodel.}

\item{MLEfit}{(Optional) An object of class \code{TFPfit} which is used for
initialization. Only used if \code{method = "bayesian"}.}
}
\description{
Checks whether model, prior and MLE fit match.
}
\keyword{internal}
