% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.checkParRestr}
\alias{.checkParRestr}
\title{Checks the given variance restrictions for consistency.}
\usage{
.checkParRestr(model, parRestr)
}
\arguments{
\item{model}{An object of class \code{NAWRUmodel} or \code{TFPmodel}.}

\item{parRestr}{list of matrices containing the parameter restrictions for the cycle,
trend, and the second observation equation (Phillips curve, CUBS equation). Each matrix
contains the lower and upper bound of the involved parameters. \code{NA} implies that no
restriction is present.}
}
\description{
Checks the given variance restrictions for consistency.
}
\keyword{internal}
