% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.checkPrior}
\alias{.checkPrior}
\title{Checks the given prior information for consistency and applicability.}
\usage{
.checkPrior(model, prior)
}
\arguments{
\item{model}{An object of class \code{TFPmodel}.}

\item{prior}{A list of matrices with parameters for the prior distribution and box
constraints.}
}
\description{
Checks the given prior information for consistency and applicability.
}
\keyword{internal}
