% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bayes.R
\name{.covAR}
\alias{.covAR}
\title{Computes the covariance of an AR(q) process.}
\usage{
.covAR(k, phi, sigma)
}
\arguments{
\item{k}{integer indicating the lag length of the covariance.}

\item{phi}{\code{q x 1} vector of parameters.}

\item{sigma}{the innovation variance of the AR(q) process.}
}
\value{
A \code{k x k} covariance matrix.
}
\description{
Computes the covariance of an AR(q) process.
}
\keyword{internal}
