% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bayes.R
\name{.covCUBS}
\alias{.covCUBS}
\title{computes the unconditional variance of the cubs equation with p lags of cubs and k
additional lags of the cycle. The cycle follows an AR process of order l.}
\usage{
.covCUBS(mu, phi, beta, sigma, phiC, sigmaC)
}
\arguments{
\item{mu}{constant.}

\item{phi}{\code{p x 1} vector of parameters for cubs lags.}

\item{beta}{\code{k x 1} vector of parameters for contemporaneous cycle and cycle lags.}

\item{sigma}{the innovation variance of the cubs process.}

\item{phiC}{\code{l x 1} vector of cycle process parameters.}

\item{sigmaC}{the innovation variance of the cycle process.}
}
\value{
A list with a two covariance matrices; the first one depicts the covariance
between the cycle and cubs and the second onf the covariance of cubs.
}
\description{
computes the unconditional variance of the cubs equation with p lags of cubs and k
additional lags of the cycle. The cycle follows an AR process of order l.
}
\keyword{internal}
