% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubs.R
\name{.cubsTa}
\alias{.cubsTa}
\title{Adjusts the frequency of cubs input series.}
\usage{
.cubsTa(tsObj, conversion, frequency)
}
\arguments{
\item{tsObj}{A time series object.}

\item{conversion}{An appropriate conversion method, i.e., \code{"average"} or \code{"sum"}.}

\item{frequency}{The frequency of the aggregated time series, i.e., \code{frequency = 4} for
quarterly and \code{frequency = 1} for annual.}
}
\description{
Adjusts the frequency of cubs input series.
}
\details{
If \code{frequency == 1}, then the values for the latest and the first year are
   averaged over the existing months/quarters. However, for the latest year, this is only
   done if the third quarter is available.
}
\keyword{internal}
