% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{.deltaMethodObs}
\alias{.deltaMethodObs}
\title{Computes standard errors of the observation equation using the delta method (for forecast).}
\usage{
.deltaMethodObs(out, nameObs, model)
}
\arguments{
\item{out}{The return object of the function \code{KFS} from the package \code{KFAS}.}

\item{nameObs}{The name of the observation equation as character.}

\item{model}{An object of class \code{NAWRUmodel}, \code{TFPmodel}, or \code{KuttnerModel}.}
}
\description{
Computes standard errors of the observation equation using the delta method (for forecast).
}
\keyword{internal}
