% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{.deltaMethodState}
\alias{.deltaMethodState}
\title{Computes standard errors of the state using the delta method.}
\usage{
.deltaMethodState(out, nameState)
}
\arguments{
\item{out}{The return object of the function \code{KFS} from the package \code{KFAS}.}

\item{nameState}{The name of the state as character.}
}
\description{
Computes standard errors of the state using the delta method.
}
\keyword{internal}
