% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfp_bayesian.R
\name{.getXYcubs}
\alias{.getXYcubs}
\title{defines Y and X in the CUBS equation.}
\usage{
.getXYcubs(stateSmoothed, model, cycleLag, cubsAR, names, trim = FALSE)
}
\arguments{
\item{stateSmoothed}{The smoothed states.}

\item{model}{An object of class TFPmodel.}

\item{cycleLag}{A non-negative integer specifying the maximum cycle lag that is included
in the CUBD equation. The default is \code{cycleLag = 0}, see details.}

\item{cubsAR}{A non-negative integer specifying the maximum CUBS lag that is included
in the CUBS equation. The default is \code{cubsAR = 0}, see details.}

\item{names}{The names of the columns.}

\item{trim}{A logical indicating whether NAs should be trimmed.}
}
\description{
defines Y and X in the CUBS equation.
}
\keyword{internal}
