% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nawru_bayesian.R
\name{.getXYpcInd}
\alias{.getXYpcInd}
\title{defines Y and X for the Phillips curve.}
\usage{
.getXYpcInd(stateSmoothed, model, cycleLag, pcIndAR, names, trim = FALSE)
}
\arguments{
\item{stateSmoothed}{The smoothed states.}

\item{model}{An object of class NAWRUmodel.}

\item{cycleLag}{A vector specifying the cycle lags that are included in the Phillip's
curve. The default is \code{cycleLag = 0}, see details.}

\item{pcIndAR}{Lag of the Phillips curve indicator.}

\item{names}{The names of the columns.}

\item{trim}{A logical indicating whether NAs should be trimmed.}
}
\description{
defines Y and X for the Phillips curve.
}
\keyword{internal}
