% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{.initializeLoc}
\alias{.initializeLoc}
\title{Initializes the location file containing default parameter constraints, among other things.}
\usage{
.initializeLoc(model)
}
\arguments{
\item{model}{A model of class \code{NAWRUmodel} or \code{TFPmodel}.}
}
\value{
A data frame containing information on each involved parameter, for instance
  its corresponding system matrix, variable names, and parameter restrictions.
}
\description{
Initializes the location file containing default parameter constraints, among other things.
}
\keyword{internal}
