% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.intervalGamma}
\alias{.intervalGamma}
\title{Computes confidence interval of Gamma distributed variable with given mean and standard
deviation.}
\usage{
.intervalGamma(mu, sd, qlower = 0.025, qupper = 1 - qlower)
}
\arguments{
\item{mu}{A \code{k x 1} vector of means.}

\item{sd}{A \code{k x 1} vector of standard deviations.}

\item{qlower}{A \code{k x 1} vector of lower quantiles.}

\item{qupper}{A \code{k x 1} vector of upper quantiles.}
}
\value{
A \code{2 x k} matrix containing the lower and upper bounds of the intervals.
}
\description{
Computes confidence interval of Gamma distributed variable with given mean and standard
deviation.
}
\keyword{internal}
