% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{.meanStd2Beta}
\alias{.meanStd2Beta}
\title{Converts the mean and standard deviation of a (possibly scaled) Beta-distributed variable
into the two shape parameters \eqn{\alpha} and \eqn{\beta}.}
\usage{
.meanStd2Beta(m, std, lb = 0, ub = 1)
}
\arguments{
\item{m}{mean parameter}

\item{std}{standard deviation}

\item{lb}{lower bound}

\item{ub}{upper bound}
}
\value{
A vector with shape parameters \eqn{\alpha} and \eqn{\beta}.
}
\description{
Converts the mean and standard deviation of a (possibly scaled) Beta-distributed variable
into the two shape parameters \eqn{\alpha} and \eqn{\beta}.
}
\keyword{internal}
