% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{.meanStd2GammasNu}
\alias{.meanStd2GammasNu}
\title{Converts the mean and standard deviation of a Gamma-distributed variable into the
parameters \eqn{s} and \eqn{\nu}.}
\usage{
.meanStd2GammasNu(m, std)
}
\arguments{
\item{m}{mean parameter}

\item{std}{standard deviation}
}
\value{
A vector with parameters \eqn{s} and \eqn{\nu}.
}
\description{
Converts the mean and standard deviation of a Gamma-distributed variable into the
parameters \eqn{s} and \eqn{\nu}.
}
\details{
The parameters \code{s} and \code{nu} are related to the regular shape
  \eqn{\alpha} and rate \eqn{\beta} parametrization in the following way:
  \eqn{\alpha = \nu / 2}
  \eqn{\beta = s / 2}
}
\keyword{internal}
