% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{.modifySSSystem}
\alias{.modifySSSystem}
\title{Modifies a an object of type \code{NAWRUmodel} or \code{TFPmodel} in case the variance
constraint for the trend is set to zero or in case a signal-to-noise ratio is specified.}
\usage{
.modifySSSystem(model, signalToNoise)
}
\arguments{
\item{model}{An object of class TFPmodel.}

\item{signalToNoise}{(Optional) signal to noise ratio. Only used if \code{method = "MLE"}.}
}
\description{
Modifies a an object of type \code{NAWRUmodel} or \code{TFPmodel} in case the variance
constraint for the trend is set to zero or in case a signal-to-noise ratio is specified.
}
\keyword{internal}
