% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubs.R
\name{.normalize}
\alias{.normalize}
\title{Normalizes a time series / vector.}
\usage{
.normalize(x)
}
\arguments{
\item{x}{E.g. a time series.}
}
\description{
Normalizes a time series / vector.
}
\keyword{internal}
