% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bayes.R
\name{.printGeweke}
\alias{.printGeweke}
\title{Prints the results of the Geweke test for the states and the parameters.}
\usage{
.printGeweke(tsl, df, alpha = 0.05)
}
\arguments{
\item{tsl}{A multiple time series list containing the Geweke test statistic for the
states.}

\item{df}{A data frame containing the Geweke test statistic for the parameters.}

\item{alpha}{The significance level used to reach a test decision, a scalar between zero
and one.}
}
\description{
Prints the results of the Geweke test for the states and the parameters.
}
\keyword{internal}
