% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{.printSSModel}
\alias{.printSSModel}
\title{Prints the model specifications.}
\usage{
.printSSModel(x, call = TRUE, check = TRUE)
}
\arguments{
\item{x}{An object of class \code{NAWRUmodel}, \code{TFPmodel}, or \code{KuttnerModel}.}

\item{call}{A logical. If \code{TRUE}, the call will be printed.}

\item{check}{A logical. If \code{TRUE}, the model class will be checked.}
}
\description{
Prints the model specifications.
}
\keyword{internal}
