% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{.printSSModelFit}
\alias{.printSSModelFit}
\title{Prints the model fit and possibly specifications.}
\usage{
.printSSModelFit(x, call = TRUE, check = TRUE, print.model = TRUE)
}
\arguments{
\item{x}{An object of class \code{NAWRUmodel}, \code{TFPmodel}, or \code{KuttnerModel}.}

\item{call}{A logical. If \code{TRUE}, the call will be printed.}

\item{check}{A logical. If \code{TRUE}, the model class will be checked.}

\item{print.model}{A logical indicating whether the model specification should be printed.}
}
\description{
Prints the model fit and possibly specifications.
}
\keyword{internal}
