% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{.priorMSd2Parameter}
\alias{.priorMSd2Parameter}
\title{Transforms the prior distribution defined by mean and standard deviation to the
appropriate input parameters.}
\usage{
.priorMSd2Parameter(
  prior,
  restr,
  namesInvGammaDistr,
  namesNormalDistr,
  namesBetaDistr
)
}
\arguments{
\item{prior}{A matrix containing the mean and standard deviation of the priors.}

\item{restr}{A matrix containing box constraints for the parameters.}

\item{namesInvGammaDistr}{A character vector specifying the names of all inverse Gamma
distributed parameters.}

\item{namesNormalDistr}{A character vector specifying the names of all normally
distributed parameters.}

\item{namesBetaDistr}{A character vector specifying the names of all Beta distributed
parameters.}
}
\description{
Transforms the prior distribution defined by mean and standard deviation to the
appropriate input parameters.
}
\keyword{internal}
