% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{.RAR2transform}
\alias{.RAR2transform}
\title{Transforms the parameters of an AR(2) process to its re-parametrized version RAR(2) and
vice versa.}
\usage{
.RAR2transform(par, phi2Atau = TRUE)
}
\arguments{
\item{par}{vector of parameters}

\item{phi2Atau}{logical indicating whether the transformation should be from AR(2) to
RAR(2) or vice versa.}
}
\value{
A vector with the transformed parameters
}
\description{
Transforms the parameters of an AR(2) process to its re-parametrized version RAR(2) and
vice versa.
}
\keyword{internal}
