% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{.SSmodelfit}
\alias{.SSmodelfit}
\title{Computes figures reagrding the model fit of the maximum likelihood estimation.}
\usage{
.SSmodelfit(out, nPar)
}
\arguments{
\item{out}{The return object of the function \code{KFS} from the package \code{KFAS}.}

\item{nPar}{A scalar specifying the number of estimated parameters.}
}
\description{
Computes figures reagrding the model fit of the maximum likelihood estimation.
}
\keyword{internal}
