% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{.SSresultsBayesian}
\alias{.SSresultsBayesian}
\title{Computes additional results of the Kalman filter and smoother for Bayesian output.}
\usage{
.SSresultsBayesian(model, HPDIprob, state, obsFitted, FUN)
}
\arguments{
\item{model}{The return object of the function \code{fitSSM} from the package \code{KFAS}.}

\item{HPDIprob}{A numeric in the interval \code{(0,1)} specifying the target probability
of the highest posterior density intervals. The default is \code{HPDIprob = 0.9}. Only
used if \code{method = "bayesian"}.}

\item{state}{An array with the smoothed state.}

\item{obsFitted}{An array with the fitted observables.}
}
\description{
Computes additional results of the Kalman filter and smoother for Bayesian output.
}
\keyword{internal}
