% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{extract_indicator_data}
\alias{extract_indicator_data}
\title{Extracts the relevant AMECO indicator data.}
\usage{
extract_indicator_data(folder)
}
\arguments{
\item{folder}{A file path with relevant files.}
}
\description{
Extracts the relevant AMECO indicator data.
}
\keyword{internal}
