% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{firstLetterUp}
\alias{firstLetterUp}
\title{Capitalizes the first letter of a string.}
\usage{
firstLetterUp(x)
}
\arguments{
\item{x}{A character.}
}
\description{
Capitalizes the first letter of a string.
}
\keyword{internal}
