% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{fit}
\alias{fit}
\title{Fit Method}
\usage{
fit(model, ...)
}
\arguments{
\item{model}{Some model.}

\item{...}{Some stuff passed on to methods.}
}
\value{
Depends on the model object, see documentation of specific methods.
}
\description{
Defines the fit method.
}
\seealso{
Other fitting methods: 
\code{\link{fit.KuttnerModel}()},
\code{\link{fit.NAWRUmodel}()},
\code{\link{fit.TFPmodel}()}
}
\concept{fitting methods}
