% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_steps.R
\name{FUNcov}
\alias{FUNcov}
\title{Computes mean and variance of the part of the posterior distribution that relies on
starting values. It then computes the density of the first p observations of Y.}
\usage{
FUNcov(Xcubs, Xc, Y, mu, phi, beta, sigma, phiC, sigmaC)
}
\arguments{
\item{Xcubs}{matrix containing lags of cubs.}

\item{Xc}{matrix containing the contemporaneous cycle and lags thereof..}

\item{Y}{a \code{Tn x 1} vector.}

\item{mu}{constant parameter.}

\item{phi}{cubs lag coeefficients.}

\item{beta}{cycle coefficients.}

\item{sigma}{innovation variance.}

\item{phiC}{cycle process parameter vector.}

\item{sigmaC}{cycle innovation variance}
}
\description{
Computes mean and variance of the part of the posterior distribution that relies on
starting values. It then computes the density of the first p observations of Y.
}
\keyword{internal}
