% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ts.R
\name{growth}
\alias{growth}
\title{Growth rate}
\usage{
growth(ts, k = 1)
}
\arguments{
\item{ts}{A time series object.}

\item{k}{An integer specifying the lag size in the growth computation. The default is
\code{k = 1}.}
}
\description{
Computes growth rates for time series.
}
\keyword{internal}
