% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{helper_fit_comparison}
\alias{helper_fit_comparison}
\title{model selection fit comparison helper function}
\usage{
helper_fit_comparison(
  fit,
  E1name,
  E1Trendname,
  E1trans = identity,
  growth,
  fitBayes
)
}
\arguments{
\item{fit}{Fitted object.}

\item{E1name}{Name of first observation equation.}

\item{E1Trendname}{Name of trend of first observation equation.}

\item{E1trans}{Transformation function for the first observation equation..}

\item{growth}{Boolean indicating whether the volatility measures should be computed for 
the growth rate of the given time series.}

\item{fitBayes}{Fitted Bayesian object.}
}
\value{
A data frame with information criteria, other goodness-of-fit measures, 
  convergence status, trend volatility measures.
}
\description{
Defines and fits models given the input parameters
}
\keyword{internal}
