% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{helper_model_comparison}
\alias{helper_model_comparison}
\title{model comparison helper function}
\usage{
helper_model_comparison(models)
}
\arguments{
\item{models}{A list of models.}
}
\value{
A data frame with model attributes
}
\description{
Gets model attributes
}
\keyword{internal}
