% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{helper_model_fit}
\alias{helper_model_fit}
\title{model selection helper function}
\usage{
helper_model_fit(FUNmodel, FUNfit, tsl, comb, poss, method, type, q, modelName)
}
\arguments{
\item{FUNmodel}{The model fitting function.}

\item{tsl}{A list of time series objects, see details.}

\item{comb}{A nested list with one model specification per list entry.}

\item{poss}{A list with the characteristics of possible models (see \code{auto.gapProd)}}

\item{method}{The estimation method. Options are maximum likelihood estimation \code{"MLE"}
and bayesian estimation \code{"bayesian"}. If \code{method = c("MLE", "bayesian")} the 
NAWRU is fitted by MLE and the TFP trend by Bayesian methods. The default is 
\code{method = "MLE"}.}

\item{type}{The variance restriction type. Possible options are \code{"basic"},
\code{"hp"}, see \code{initializeRestr}. The default is \code{type = "hp"}.}

\item{q}{Quantile for the Inverse Gamma distribution (only used if \code{type = "hp"}), 
see \code{initializeRestr}. The default is \code{q = 0.01}.}

\item{modelName}{Name of the model, i.e., NAWRU or TFP.}
}
\value{
A nested list with the models and fitted objects.
}
\description{
Defines and fits models given the input parameters
}
\keyword{internal}
