% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{inference}
\alias{inference}
\title{Computes standard errors, t-statistics, and p-values for the estimated state space parameters
using the delta method.}
\usage{
inference(parOptim, hessian, loc)
}
\arguments{
\item{parOptim}{The vector of optimized parameters, without transformations.}

\item{hessian}{The hessian from the optimization.}

\item{loc}{A \code{3 x n} array where \code{n} is the number of optimized parameters. The array
contains information on the estimated parameters's name (\code{loc[1, ]}), its location
(\code{loc[2, ]}) and possible parameter constraints (\code{loc[3, ]}).}
}
\description{
Computes standard errors, t-statistics, and p-values for the estimated state space parameters
using the delta method.
}
\keyword{internal}
