% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{initializeRestr}
\alias{initializeRestr}
\title{Initialization of parameter restrictions}
\usage{
initializeRestr(model, type = "basic", lambda = NULL)
}
\arguments{
\item{model}{An object of class \code{NAWRUmodel}, \code{TFPmodel}, or \code{KuttnerModel}.}

\item{type}{The variance restriction type. Possible options are \code{"basic"},
\code{"hp"}, see details. The default is \code{type = "basic"}.}

\item{lambda}{The smoothing constant for the HP-filter if \code{type = "hp"}.}
}
\value{
A list of three matrices containing the parameter restrictions for the cycle,
  trend, and the second observation equation. Each matrix contains the lower and upper
  bound of the involved parameters. \code{NA} implies that no
  restriction is present.
}
\description{
Initializes parameter restrictions for objects of class \code{NAWRUmodel},
\code{TFPmodel}, or \code{KuttnerModel}.
}
