% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfp.R
\name{is.TFPfit}
\alias{is.TFPfit}
\title{\code{TFPfit} object check}
\usage{
is.TFPfit(object, return.logical = FALSE)
}
\arguments{
\item{object}{An object to be tested.}

\item{return.logical}{If \code{return.logical = FALSE} (default), an error message is printed
if the object is not of class \code{TFPfit}. If \code{return.logical = TRUE}, a logical
value is returned.}
}
\value{
A logical value or nothing, depending on the value of \code{return.logical}.
}
\description{
Tests whether the input object is a valid object of class \code{TFPfit}.
}
\keyword{internal}
