% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matmult3d}
\alias{matmult3d}
\title{Capitalizes the first letter of a string.}
\usage{
matmult3d(a, b)
}
\arguments{
\item{a}{A matrix.}
}
\description{
Capitalizes the first letter of a string.
}
\keyword{internal}
