% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ts.R
\name{operTsLists}
\alias{operTsLists}
\title{Performs a mathematical operation to the ts elements of two lists with the same names}
\usage{
operTsLists(x, y, operator)
}
\arguments{
\item{x}{A list.}

\item{y}{A list.}

\item{operator}{mathematical operator in quotation marks, e.g., "*".}
}
\description{
Performs a mathematical operation to the ts elements of two lists with the same names
}
\keyword{internal}
