% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfp.R
\name{plot.TFPfit}
\alias{plot.TFPfit}
\title{Plots for a \code{TFPfit} object}
\usage{
\method{plot}{TFPfit}(
  x,
  alpha = 0.05,
  bounds = TRUE,
  path = NULL,
  combine = TRUE,
  prefix = NULL,
  posterior = FALSE,
  device = "png",
  width = 10,
  height = 3,
  ...
)
}
\arguments{
\item{x}{An object of class \code{TFPfit}.}

\item{alpha}{The significance level for the TFP trend (\code{alpha in [0,1]}). Only used if
\code{bounds = TRUE}.}

\item{bounds}{A logical indicating whether significance intervals should be plotted around
tfp growth. The default is \code{bounds = TRUE}.}

\item{path}{An optional file path. If specified, the plots will be saved using the format
in \code{device} under the given path.}

\item{combine}{A logical indicating whether the diagnostic plots should be combined or not,
the default is \code{TRUE}.}

\item{prefix}{An optional character string to be added to the names of the plots in case
\code{path} is specified.}

\item{posterior}{A logical indicating whether posterior diagnostics should be plotted. The
default is \code{FALSE}. Only applied in the case of bayesian estimation.}

\item{device}{Device passed on to \code{ggplot} for plot saving. Options are eps", "ps",
"tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf".}

\item{width}{The plot width in case of printing.}

\item{height}{The plot height in case of printing.}

\item{...}{Ignored.}
}
\description{
Plots the TFP trend and the CUBS equation and gives diagnostic plots based on
standardized residuals for objects of class \code{TFPfit}.
}
