% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gibbs_output}
\alias{plot_gibbs_output}
\title{Plots the diagnostic plots of the posterior distribution.}
\usage{
plot_gibbs_output(
  path = NULL,
  draws,
  parName,
  burnin,
  mu = NULL,
  prec = NULL,
  shape = NULL,
  scale = NULL,
  shape1 = NULL,
  shape2 = NULL,
  ub = NULL,
  lb = NULL,
  prefix = NULL,
  device,
  width,
  height
)
}
\arguments{
\item{path}{An optional file path. If specified, the plots will be saved using the format
in \code{device} under the given path.}

\item{draws}{the number of draws.}

\item{parName}{the name of the parameter.}

\item{burnin}{length of thw burnin phase.}

\item{mu}{mean for normal distribution}

\item{prec}{precision for normal distribution}

\item{shape}{shape for Gamma distribution}

\item{scale}{scale for Gamma distribution}

\item{shape1}{shape1 for Beta distribution}

\item{shape2}{shape2 for Beta distribution}

\item{ub}{upper bound}

\item{lb}{lower bound}

\item{prefix}{An optional character string to be added to the names of the plots in case
\code{path} is specified.}

\item{device}{Device passed on to \code{ggplot} for plot saving. Options are eps", "ps",
"tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf".}

\item{width}{The plot width in case of printing.}

\item{height}{The plot height in case of printing.}
}
\description{
Plots the diagnostic plots of the posterior distribution.
}
\keyword{internal}
