% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict}
\alias{predict}
\title{Prediction Method}
\usage{
predict(object, ...)
}
\arguments{
\item{object}{Some object.}

\item{...}{Some stuff passed on to methods.}
}
\description{
Defines the predict method as it does not seem to exist.
}
\keyword{internal}
