% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predictBayes}
\alias{predictBayes}
\title{Predictions for Bayesian estimation}
\usage{
predictBayes(fit, n.ahead = 10, exogenous = "mean", returnFit = TRUE)
}
\description{
Computes predictions for an object of class \code{fitNAWRU, fitTFP}
estimated via Bayesian methods.
}
\keyword{internal}
