% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predictMLE}
\alias{predictMLE}
\title{Predictions for MLE}
\usage{
predictMLE(fit, n.ahead = 10, exogenous = "mean", returnFit = TRUE)
}
\description{
Computes predictions for an object of class \code{fitNAWRU, fitTFP, fitKuttner}
estimated via MLE.
}
\keyword{internal}
