% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.R
\name{print.gap}
\alias{print.gap}
\title{Print \code{gap} object}
\usage{
\method{print}{gap}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gap}.}

\item{...}{Ignored.}
}
\description{
Prints the model specifications of an object of class \code{gap}.
}
