% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuttner.R
\name{print.KuttnerFit}
\alias{print.KuttnerFit}
\title{Print \code{KuttnerFit} object}
\usage{
\method{print}{KuttnerFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{KuttnerFit}.}

\item{...}{Ignored.}
}
\value{
No return value, results are printed.
}
\description{
Prints the model specifications and the estimation results of an object of
  class \code{KuttnerFit}.
}
