% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuttner.R
\name{print.KuttnerModel}
\alias{print.KuttnerModel}
\title{Print \code{KuttnerModel} object}
\usage{
\method{print}{KuttnerModel}(x, call = TRUE, check = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{KuttnerModel}.}

\item{call}{A logical. If \code{TRUE}, the call will be printed.}

\item{check}{A logical. If \code{TRUE}, the model class will be checked.}

\item{...}{Ignored.}
}
\description{
Prints the model specifications of an object of class \code{KuttnerModel}.
}
