% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfp.R
\name{print.TFPfit}
\alias{print.TFPfit}
\title{Print \code{TFPfit} object}
\usage{
\method{print}{TFPfit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{TFPfit}.}

\item{...}{Ignored.}
}
\value{
No return value, results are printed.
}
\description{
Prints the model specifications and the estimation results of an object of class \code{TFPfit}.
}
