% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfp.R
\name{print.TFPmodel}
\alias{print.TFPmodel}
\title{Print \code{TFPmodel} object}
\usage{
\method{print}{TFPmodel}(x, call = TRUE, check = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{TFPmodel}.}

\item{call}{A logical. If \code{TRUE}, the call will be printed.}

\item{check}{A logical. If \code{TRUE}, the model class will be checked.}

\item{...}{Ignored.}
}
\value{
No return value, model information is printed.
}
\description{
Prints the model specifications of an object of class \code{TFPmodel}.
}
