% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{trendOptim}
\alias{trendOptim}
\title{Find suitable trend specification}
\usage{
trendOptim(x, opt = c("RW1", "RW2", "DT"))
}
\arguments{
\item{x}{A time series.}

\item{opt}{A character vector with the trend models to be tested. The default is 
\code{opt = c("RW1", "RW2", "DT")}.}
}
\value{
A character string with the chosen trend model.
}
\description{
Finds the most suitable trend model according to the BIC.
}
\keyword{internal}
