% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{trendVolaMeasures}
\alias{trendVolaMeasures}
\title{Trend volatility measures}
\usage{
trendVolaMeasures(tsOriginal, tsTrend, growth = FALSE)
}
\arguments{
\item{tsOriginal}{The original time series.}

\item{tsTrend}{The trend time series.}

\item{growth}{Boolean indicating whether the volatility measures should be computed for 
the growth rate of the given time series.}
}
\value{
A list containing the different measures.
}
\description{
Computes trend volatility measures.
}
\keyword{internal}
