% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{initializeExo}
\alias{initializeExo}
\title{Initialization of exogenous variables.}
\usage{
initializeExo(maxDiff = 1, maxLag = 1, varNames)
}
\arguments{
\item{maxDiff}{An integer specifying the maximal difference order.}

\item{maxLag}{An integer specifying the maximal lag order.}

\item{varNames}{A \code{(k x 1)} character vector containing the names of the exogenous
variables.}
}
\value{
An array of size \code{(2, k, max(maxDiff, maxLag) + 1)}. The first row specifies
  the difference order and the second the lag order.
}
\description{
Initializes the transformation of the exogenous variables.
}
