% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predictBayes}
\alias{predictBayes}
\title{Predictions for Bayesian estimation}
\usage{
predictBayes(fit, n.ahead = 10, exogenous = "mean", returnFit = TRUE)
}
\arguments{
\item{fit}{An object of class \code{NAWRUfit}, \code{TFPfit}, or \code{KuttnerFit}.}

\item{n.ahead}{An integer specifying the prediction horizon.}

\item{exogenous}{A character string specifying the computation of exogenous variables 
included in the model (if applicable). Valid options are \code{exogenous = "mean"} and 
\code{exogenous = "last".}}

\item{returnFit}{A logical. If \code{TRUE}, an object of the same class as \code{fit}
where the list entry \code{tsl} is replaced. If \code{FALSE}, only the new time series 
list is returned.}
}
\description{
Computes predictions for an object of class \code{fitNAWRU, fitTFP}
estimated via Bayesian methods.
}
\keyword{internal}
