% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nawru.R
\name{print.NAWRUfit}
\alias{print.NAWRUfit}
\title{Print \code{NAWRUfit} object}
\usage{
\method{print}{NAWRUfit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{NAWRUfit}.}

\item{...}{Ignored.}
}
\value{
No return value, results are printed.
}
\description{
Prints the model specifications and the estimation results of an object of class \code{NAWRUfit}.
}
