% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{trendVolaMeasures}
\alias{trendVolaMeasures}
\title{Trend volatility measures}
\usage{
trendVolaMeasures(tsOriginal, tsTrend, nDiff)
}
\arguments{
\item{tsOriginal}{The original time series.}

\item{tsTrend}{The trend time series.}

\item{nDiff}{Integer indicating the order of differencing applied to the input series.}
}
\value{
A list containing the different measures.
}
\description{
Computes trend volatility measures.
}
\keyword{internal}
